Add some helpers to simplify cross-version/cross-platform modules (#2406)
This PR attempts to minimize the boilerplate of defining cross-version
cross-platform Scala modules.

Currently, cross-version cross-platform modules involves a lot annoying
fiddling with overriding `millSourcePath` and `sources` and
`artifactName` and passing around `crossScalaVersion` and all that.
Simple `Cross`/`CrossScalaModule`s get some of the plumbing handled
behind the scenes, but anything more complicated and the plumbing spills
out again into plain view. And although `Cross`/`CrossScalaModule`s have
less boilerplate, it's still enough boilerplate you don't want to go
around defining them over and over

This PR tries to hide all of them behind helper traits
`CrossScalaModule.Wrapper` and `PlatformScalaModule`, so your build file
only has the logic you care about and none of the plumbing.

## Major changes

1. We replace the `artifactName` definition, so that instead of reading
from `millModuleSegments` directly, it reads from `artifactNameParts:
Seq[String]`. This makes it easier for us to splice the artifact name
parts to remove unwanted segments without needing to do fragile string
manipulation

2. Introduce `CrossScalaModule.Wrapper`. This can be used when the
`ScalaModule`s you are defining are not the direct children of the
`Cross` module, but are nested somewhere inside of it.
`CrossScalaModule.Wrapper` shades the `CrossScalaModule` trait, so that
any `CrossScalaModule` defined within its body would have the
`crossScalaVersion` automatically wired up and the `artifactNameParts`
properly adjusted to remove the cross segment in the middle of the
artifact name

3. Introduce `PlatformScalaModule`. This is analogous to
`CrossScalaModule`, in that it adjusts the
`millSourcePath`/`artifactNameParts` to drop the last segment, and
adjusts `sources` to include the `-{jvm,js,native}` suffix

## Tests

The changes are largely covered by existing tests, including the
`example/` tests which were adjusted to make use of the new traits

## Notes

1. We cannot make the platform-specific sub-modules into `Cross` modules
like we do with the version-specific sub-modules, since the
platform-specific submodules have pretty different types/targets defined
on them and `Cross` modules must all be of the same type


2. `6-cross-platform-publishing`, is a lot better than before, but there
is still a significant amount of boilerplate involved, especially the
lines around `object wrapper`. Not sure if there are other things we can
do to make it even easier to define these cross-version/platform modules