Pre compile zinc compiler interface for common Scala versions (#2424)
While we still the to keep the dynamic codepath necessary to compile
stuff at runtime, at least for the common case of existing Scala
versions we can pre-compile it, improving performance and removing some
of the confusing log messages about compiling stuff that the user didn't
write. This only affects the first load, but first impressions matter.
Doesn't benefit Scala 3, but there's half the community is still on
Scala 2, and Mill's own `build.sc` is still on Scala 2

Before, we can see that a clean `run` ends up compiling the compiler
interface twice (once to compile `build.sc`, once to compile the user's
`ScalaModule`) taking ~14s on my laptop

```
Compiling compiler interface...
[info] compiling 1 Scala source to /Users/lihaoyi/Github/mill/example/basic/1-hello-world/out/mill-build/compile.dest/classes ...
[info] done compiling
[34/48] compile
Compiling compiler interface...
8 warnings
[info] compiling 1 Scala source to /Users/lihaoyi/Github/mill/example/basic/1-hello-world/out/compile.dest/classes ...
[warn] 1 deprecation (since 2.13.0); re-run with -deprecation for details
[warn] one warning found
[info] done compiling
[48/48] run
<h1>hello</h1>
/Users/lihaoyi/Github/mill/target/mill-release -i run --text hello  47.58s user 1.70s system 343% cpu 14.345 total
```

After, we compile the compiler interface zero times, with a clean `run`
taking ~7s on my laptop

```
[build.sc] [40/47] compile
[info] compiling 1 Scala source to /Users/lihaoyi/Github/mill/example/basic/1-hello-world/out/mill-build/compile.dest/classes ...
[info] done compiling
[34/48] compile
[info] compiling 1 Scala source to /Users/lihaoyi/Github/mill/example/basic/1-hello-world/out/compile.dest/classes ...
[warn] 1 deprecation (since 2.13.0); re-run with -deprecation for details
[warn] one warning found
[info] done compiling
[48/48] run
<h1>hello</h1>
/Users/lihaoyi/Github/mill/target/mill-release -i run --text hello  19.27s user 0.85s system 297% cpu 6.768 total
```

## Implementation

I added cross modules `bridge[_]` for all supported Scala versions.
These more or less follow what `ZincWorkerModule#scalaCompilerBridgeJar`
already does on-the-fly: download the respective source jar from Maven
Central and compile it using the respective Scala version. Note that the
`META-INF` metadata is necessary for Zinc to properly pick these up, and
so I manually copy the folder from the unzipped source jar into the
`resources` so it can get propagated to the final jar.

To avoid the slowness of compiling every bridge version during local
development, we separate the `bridge[_]` publishing from the rest of the
build, only enabled via the `MILL_BUILD_COMPILER_BRIDGES=true`
environment variable, and published under a separate version. We expect
to bump the `bridge[_]` version rarely, and want to avoid redundantly
re-publishing them every Mill release.

## Testing

Most of the existing unit tests use the three scala versions that we
pre-compile in development mode, though some of them are left on other
versions for legacy reasons e.g. they depend on a specific version of
Scala to be compatible with a specific version of semanticDB or
scala-native, and those continue to use the old
download-compile-on-the-fly code path.

I also added a new test case
`mill.scalalib.HelloWorldTests.compile.nonPreCompiledBridge` that
together with `.fromScratch` specifically exercises the code paths for
compiled/not-compiled compiler bridges, asserting that the compiler
bridge is compiled in the versions it should be compiled for (i.e. those
that are not pre-compiled) and not compiled for versions it should not
be compiled for (i.e. those that *are* pre-compiled)