enforce stdout/stderr contents in `.local` example/integration tests (#2435)
This should allow us to do debugging or iteration on example/integration
in `.local` mode, which is a lot faster than `.fork` or `.server`

Even after https://github.com/com-lihaoyi/mill/pull/2425 landed, the
`.local` integration tests are still much faster to run than `.fork` and
`.server`, with publishing taking ~20s down from ~60s before which is
still a very annoying wait. Better if we can avoid that, using `.local`
in the common development workflow and leaving `.fork` and `.server` for
CI to catch rare edge-case bugs

We basically needed to be a bit more robust in spawning subprocesses,
allowing the `InputPumper` machinery to work for any redirected streams
rather than being hardcoded to work with the `PipedInputStream` that
`MillServerMain` uses.

I had to fix an issue in `ScalaJSModule#run` to make
`example.web[3-scalajs-module].local.test` pass with this additional
enforcement. It appears we were not properly pumping the JSEnv
subprocess stdout/stderr. This would cause the outputs to disappear when
run repeatedly in client-server mode, since the output would be
inherited by first MillClient process that spawned the server, and not
subsequent clients that simply connect to it

---------

Co-authored-by: Tobias Roeser <le.petit.fou@web.de>