Consolidate Docsite/Examples/Scaladoc/Giter8 (#2448)
This PR aims to better integrate the four disparate sources of
documentation we maintain in the Mill repo:

1. The docsite in `docs/`: nice pedagogical snippets and explanations
written over time, but not tested and generally buggy/bitrotted over
time

2. The examples and integration tests in `example/` and `integration/`,
which was originally mostly integration tests but now have a good number
of `examples/` with good comments and a mostly readable bash-like DSL
test harness

3. The Scaladoc, scattered throughout the codebase

4. The `mill init` Git8r templates, not strictly in this repo, but we
maintain the functionality to pull stuff from elsewhere

Currently, all four formats have kind of mixed concerns: they all have
tutorial-ish stuff, they all have reference-ish stuff, they all have
code snippets here and there.

This PR aims to do the following:

1. Re-use as much of `example/` in `docs/` as possible
* We already are parsing the comments to extract the assertions we need
to run, so it's easy to also take those parsed comments and generate
nice AsciiDoc

3. Move as many code snippets from `docs/` into `examples/` as possible.
* Since now we can easily `include::` the examples in our doc-site, the
cost of doing this is low.
* The benefit is that now the doc-site snippets become tested for
correctness, and easily download-able for the user to experiment with
themselves

5. Enrich the Scaladoc to better serve as reference documentation for
the various classes.
* In particular, we want to avoid methods or classes showing up in the
Scaladoc un-documented
* We want to link to the Scaladoc prominently from the doc-site in
places where a reader may want a reference they can look up.
* We probably can move some of the more reference-y docs from the
doc-site purely into the Scaladoc, e.g. the page on `Task Context API`,
or some of the `Built-in commands` stuff in the `Getting Started`

7. Integrate the examples into `mill init`
* The example repos provide excellent templates for users to get started
with a wide range of projects, so we should leverage them as much as
possible to help users bootstrap their Mill projects.
    * This can be in addition to the existing giter8 support

It's not entirely clear whether we should write `build.sc` files and
parse them to generate the `.adoc`, or write `.adoc` files and parse
them to generate the `build.sc`. I chose the former, but if we change
our mind we can use the parser we already have to auto-transform them to
the opposite format if necessary.


Progress reviewing docsite pages

- [x] Introduction to Mill
- [x] Installation and IDE Support
- [x] Getting Started --> move to `example/` and Scalaoc
- [x] Learn by Example
- [x] Scala Module Configuration
- [ ] Library Dependencies in Mill
- [ ] The Output Directory
- [x] Tasks --> move to `example/`
- [x] Task Context API --> move to Scaladoc
- [x] Modules --> move to `example/` and Scalaoc
- [x] Cross Builds
- [x] Web Development Examples
- [ ] Extending Mill
- [ ] Using Plugins
- [x] Mill Internals
- [x] Contrib Plugins
- [x] Third-Party Plugins
- [x] External References
- [x] Mill Scaladoc

---------

Co-authored-by: Tobias Roeser <le.petit.fou@web.de>

Pull request: https://github.com/com-lihaoyi/mill/pull/2448