Cleanup mill.eval (#2498)
# Major Changes

1. Split out a public `Evaluator` interface from the private
`EvaluatorImpl`
1. This allows us to write the private code cleanly without mixing it up
with verbose binary-compatibility workarounds
2. It also significantly limits the APIs we make public; now public APIs
on `Evaluator` are opt-in, rather than opt-out via the `private`
keyword. A lot of things that were previously public but probably don't
need to be public are now private

2. Combine `sequentialEvaluate` and `parallelEvaluate` into a single
implementation, with the only difference between the two being the usage
of `ExecutionContexts.RunNow` v.s. `ExecutionContexts.ThreadPool`. This
cleans up a significant amount of copy-pasty code between the two

3. Changed `Evaluator.Terminal` from a type alias for `Either` to a
`sealed trait` hierarchy. This significantly reduces the boilerplate we
have boxing/unboxing things in `Left` and `Right`, with the new names
`Labelled` and `Task` being much more semantically meaningful.

# Minor Changes

1. `Evaluator.evalOrThrow(evaluator)` is now `evaluator.evalOrThrow`, to
avoid a circular dependency between `Evaluator` and `EvaluatorImpl`

2. `private`ized some other stuff in `mill.eval`: `Graph`, `Tarjans`,
etc.

3. Break out `loadExistingWorker` and `loadCachedJson` from
`evaluateGroupCached` to try and reduce the nesting