Enable `inspect`, `path`, etc. on `T.command`s without needing to pass in args in common cases (#2500)
In general, *most* commands do not depend on their arguments to
instantiate the `mill.Command` object, and only depend on their
arguments to actually execute. For that common case, we should be able
to pass in `null`s for the arguments, which would allow us to do
`inspect`, `path`, `plan`, etc. without needing to tediously pass in
dummy arguments at the command line.

We add a `nullCommandDefaults` flag to calls to `Resolve*.resolve` in
`MainModule.scala`, and make it override the MainArgs `ArgSig`s to set
all default values to `null` when passed (`T.task(null)` for `Task[_]`
params, primitive params it gets casted to `0` for numbers, `false` for
booleans, etc.). This allows us to instantiate the `Command` object and
analyze the task graph, even if we aren't able to actually execute it

This PR also removes support for `SelectMode.Single`, since it is
largely superseded by `SelectMode.Separated`. Looking at the various use
sites of `SelectMode.Single`, I don't think avoiding the need to escape
`\\+` in edge cases is worth the additional `SelectMode` when
`Separated` is close enough

Would fix https://github.com/com-lihaoyi/mill/issues/503. Fixes
https://github.com/com-lihaoyi/mill/issues/937 in the common case where
the commands in question instantiate their `mill.Command` task graph and
metadata without depending on the input arguments; the scenario where we
`if`/`else` on the input arguments and return different `T.command`s in
each case is not resolved by this PR, but should be a pretty rare
scenario so better to fix the common case first