Improve VSCode/BSP integration (#2514)
https://github.com/com-lihaoyi/mill/pull/2415 only tested the Bloop
codepath, and left the BSP code path broken. This fixes it

Tested manually by setting `switch build server` in VSCode to BSP and
then clicking around in the following examples:

1. `example/basic/1-simple-scala`
2. `example/misc/3-import-file-ivy`
3. `example/misc/4-mill-build-folder`
 
The following navigations work:

1. between `foo` and `foo.test`
2. between `Foo` and the Java and Scala std libs
3. between `Foo` and Scalatags
4. between `build.sc` and Mill library code
5. between `build.sc` and Scalatags
6. between `build.sc` and `mill-build/src/`
7. between `mill-build/build.sc` and Mill library code
8. between `mill-build/build.sc` and Java and Scala std libs

I re-used some code from `GenIdeaImpl` to gather the source jars of mill
bundled modules in `MillBuildServer`.

TBH the code in both `MillBuildServer` and `GenIdeaImpl` is a mess and
could use some housekeeping, and `MillBuildServer` could also use some
unit/e2e tests to verify basic the basic shape of that
`workspaceBuildTargets`, `buildTargetSources`,
`buildTargetDependencySources` and others return in a few simple example
cases. But that can probably come in a follow up