Make mill-build folder opt in via `import $mill.build` (#2527)
Fixes https://github.com/com-lihaoyi/mill/issues/2461

For now it remains hard-coded as `mill-build/`, for simplicity. Given
the `mill-` prefix, it's unlikely to collide with user-defined folders.
Making it configurable would open another can of worms where the `import
$meta.foo` can now collide with an `object foo`, which is very
unintuitive.

For now, we just parse the script files one additional time in
`MillBuildBootstrap.scala`. This is a bit wasteful, but is probably fast
enough for now, and we're already pretty sloppy parsing everything twice
in `MillBuildRootModule#scriptSources` and
`MillBuildRootModule#parseBuildFiles`, so parsing things three times
isn't the end of the world. We can look into optimizing it in future if
necessary