Use `showNamed` output format for `show` with more than one item (#2532)
The old array-based output is pretty useless since it doesn't tell you
what targets each output actually corresponds to. For humans, `show` is
now probably what they want 99% of the time, and it simplifies having to
remember to use `show` or `showNamed` depending on how many targets
their query returns.

We leave `showNamed` for the "always print out dictionary" behavior,
which might be useful for external tools who would not appreciate the
output format flip-flopping between a string and a dictionary depending
on the target count.