Make `mill.define.Module` a trait to allow abstract/virtual module `lazy val`s to be `trait`s rather than `class`es (#2536)
The basic problem is that when `Module` was a class, and you had `trait
FooModule extends Module`, Java reflection would not see that `FooModule
<: Module`, and thus `Reflect.reflectNestedObjects0[Module]` which
relies on java reflect would not detect `def foo: FooModule` as a
module. This is because `trait extends class` is a Scala compiler
concept, and does not exist in the JVM bytecode.

However, we do need `Module` to be a class, because we rely on implicit
parameters to propagate context down the module tree. `trait`s cannot
take implicit parameters

The solution here is to make `Module` a trait, and make it inherit from
a `Module.BaseClass` class that takes the implicit parameter and passes
it to `Module`. Thus whenever we do reflection to find `FooModule <:
Module`, it behaves expected, and we are still able to receive implicit
parameters via `Module.BaseClass`

One side effect is that we need a way to mark things like
`JavaModule#zincWorker` as _not_ a child module. Previously, we skipped
such modules during target resolution due to the buggy behavior above.
Now, such modules get picked up, which is not what we want: in the
common case, `zincWorker` is just a reference to the shared external
module `mill.scalalib.ZincWorkerModule`, and is not a child module of
each individual `JavaModule`. To do this, I introduce a
`mill.define.ModuleRef` wrapper that simply provides a thin layer of
indirection, letting us refer to the module via `zincWorker()` but
blocking identification as a child-module since `ModuleRef` is not a
sub-type of `Module`.

Tested via the `mill.resolve.ResolveTests.overridenModule` tests, which
now pass when the abstract/overriden module is a trait, whereas
previously they required that it be a class. Should also fix the
flakiness we've been seeing in
`mill.resolve.ResolveTests.overridenModule.0`.