Update thirdparty tests, make subprocess testrunner run user code in root classloader (#2561)
1. Move `thirdparty` from `integration/` to `example/`, and simplify
them down to `acyclic`, `fansi`, and `jimfs`. We can add more complex
builds later if we want, but for now just having some realistic builds
that are not bitrotted and old is already an improvement over what we
have now.

2. Flip the sub-process `TestRunner` classloaders on its head: rather
than runner `TestRunner.scala` in the boot classloader and user code in
a sub-classloader, we run user code in the boot classloader and
`TestRunner.scala` in a sub classloader. This is necessary to make JimFS
work, and would probably fix
https://github.com/com-lihaoyi/mill/issues/716 and related issues.

3. Broke up and cleaned up `TestRunner.scala`:
1. Separated `TestRunner` and `TestRunnerMain0`, the two entrypoints,
from `TestRunnerUtils`
2. Replaced all the complicated logic for converting logic to/from CLI
flags with JSON written to a file by `TestModule` and read from the file
by `TestRunner`
3. Replaced `ClosableIterator` with `geny.Generator`, which also
provides guaranteed closing after iteration

4. Marked most of `testrunner` as `@internal`. Most of these could be
`private[mill]`, but I think we don't have a clear idea of how this
stuff is meant to be used yet, and it's plausible that other people may
want to interact with the testrunner in their own custom test suites. I
think `@internal` is better suited for now until we either (a) know
better what a stable API looks like or (b) decide that this stuff should
not be used externally.

We have much more control over `TestRunner.scala` than we do over user
code, so we can make sure that `TestRunner.scala` works when running in
a semi-isolated classloader, whereas for user code we can offer no such
guarantees. We can't do anything for `.testLocal` that runs in-process,
but at least for `.test` we can try to offer them a clean classloader
environment.

The classloader for user code is not 100% clean: we still have a single
class `mill.testrunner.entrypoint.TestRunnerMain` added within it.
That's probably OK, since no matter what we do w.r.t. shading and stuff
we still need somewhere to put our own main method to bootstrap the
`mill.testrunner` classloader.

The testrunner changes are covered by existing unit and
integration/example tests, and the new "user code runs in clean
classloader" logic is covered by `example.thirdparty[jimfs].local`.