Make empty selector message more helpful (#2610)
Regarding #2601. Currently, if no arguments are given to mill, it says
`Selector cannot be empty`. It's not immediately clear what that's
supposed to mean, so I figured it might be more helpful to print the
help message instead in that case (some other tools do that). In
addition to this, I modified the empty selector message to say
``Target selector must not be empty. Try `mill resolve _` to see what's
available.``.

Pull request: https://github.com/com-lihaoyi/mill/pull/2610