Differentiate by  Mill version when selecting a server worker (#2639)
In older Mill versions, we tried to detect when the Mill version
changed. In that case, we stopped the server process and started a new
one. This server stopping worked unrelieble and is currently not working
at all.

This pragmatic change simply does not try to re-use a mill server
process when the Mill version does not match. Instead, we only choose
between instances with the exact version or start a new instance. This
is identical to the beavior, when we run with a differnet JVM.

Resource wise, we potentially increase the amount of running processes,
but we do this anyways in other situations. Also, idle Mill servers stop
themselves after a some time span.

Pull request: https://github.com/com-lihaoyi/mill/pull/2639