use bspCompileClassesPath in bspTransitiveCompileClasspath (#2646)
Hi! This is my first contribution, so please let me know if there is
anything for me to do.

As far as I understand this is OK, since `bspCompileClasspath` mentions
it tries not to trigger compilation, so the transitive version may also
try not to trigger compilation of the module dependencies.

This would be helpful to us, since we currently cannot load projects in
intellij via bsp if there is a compilation error in any module other
than the topmost module. On the surface this seems similar to the issue
in #2484 (failing on `buildTarget/scalacOptions` due to an error within
`_.compile`).

Pull request: https://github.com/com-lihaoyi/mill/pull/2646