Fix init command and support running Mill without existing `build.sc`
I first added an integration test for the init command to reproduce the issues and make sure the feature now works.

Instead of failing in case there is no `build.sc`, we continue with an empty setup.

Only when we discover an error afterwards, we add a message about the missing `build.sc` to the underlying error message.