Fix line numbers in errors for scripts starting with leading comments or whitespace (#2686)
This is basically a port of the following fix in the Ammonite repo
https://github.com/com-lihaoyi/Ammonite/pull/1361. There were some
changes when the logic was first ported from Ammonite to Mill, resulting
in a much smaller code change than was necessary in the Ammonite repo

There are two related follow up PRs that AFAICT do not apply to Mill:
https://github.com/com-lihaoyi/Ammonite/pull/1363
and https://github.com/com-lihaoyi/Ammonite/pull/1369. Both have to do
with the special handling of multi-statement top-level-blocks `{...}`,
that is a thing in the Ammonite REPL (to allow multiple top-level
declarations to be entered into the REPL before submitting it) but
doesn't apply to Mill scripts.

I have updated the integration tests to include a bunch of random
comments and imports to reproduce the problem causing the tests to fail,
and verified the fix causes the tests to pass. Also this caused a
previously passing test to fail (`2-custom-build-logic`), which on
inspection appears like it was asserting the wrong thing, so I fixed the
assert

Fixes https://github.com/com-lihaoyi/mill/issues/2681